playing with intellij
git-svn-id: https://svn.code.sf.net/p/yarfraw/code/yarfraw/yarfraw@98 fa792110-9d37-0410-8bf1-8cf09365c9f7

+++ src/main/java/yarfraw/core/datamodel/Cloud.java
@@ -78,7 +78,7 @@ public class Cloud extends AbstractBaseObject{
     }
     
     try {
-      if(_port != null && (_port.intValue() < 0 || _port.intValue() > 65535)){
+      if(_port != null && (_port < 0 || _port > 65535)){
         throw new ValidationException("Cloud: Invalid port number");
       }
     }
